indexing: rename index.py to textify.py
The module now holds only the textifier; `index` was a leftover catch-all
name. Pure rename + repointed imports.
